Update ember version to 1.13.x to support `hasBlock`
We have to stick the minimum version of ember for this addon to 1.13.x
since 1.12.x was the last version of Ember that supports checking for a
block like this:
```hbs
        {{#if template}}
              ...
        {{/if}}
```

Template is deprecated and not used in 1.13. So in order to be
compatiable with newer Ember versions we need to use `hasBlock` instead
of `template`:
 ```hbs
        {{#if hasBlock}}
              ...
        {{/if}}
```

I also updated mocha to the latest versions so ember 2.0 builds will
pass.