Add the ability to use a blockless form
Unfortunately this ups the minium version of Ember back up to 1.13 since
we have to check for a block.

Also caught a bug where you couldn't have more than 1 of these
components on a page due to them having the same ID. Short sighted, I
know. It's fixed now with a randomly generated ID.