Add the ability to reset the file input
When the `action` is called we now send a second argument that is a
function. This function is a reset function that you can invoke after
you've finished working with your FileList blob.

The reset function just resets the value of the file input to an empty
string, allowing you to select the same file as many times as you want.